backend/requirements.txt
gunicorn
Flask
SQLAlchemy
requests
python-dotenv
psycopg2-binary
